+++ CHANGELOG.md
@@ -13,6 +13,8 @@
  - Make `Resizable` have a `gameSize` property instead of `size`
  - Fix bug with CombinedEffect inside SequenceEffect
  - Fix wrong end angle for relative rotational effects
+ - Use a list of Vector2 for Move effect to open up for more advanced move effects
+ - Generalize effects api to include all components
 
 ## [next]
  - Fix spriteAsWidget deprecation message